Enhance TripPlanner app by adding a checklist feature for trips, allowing users to create, view, and update tasks. Implemented ItemRepository and ItemDao for managing checklist items. Updated TripDetailActivity to display tasks with checkboxes and ensure default tasks are created if none exist. Adjusted SplashActivity to enforce login on app start. Updated README.md to reflect new features and security improvements.

+++ app/src/main/java/com/example/tripplanner/data/local/dao/ItemDao.java
@@ -26,5 +26,5 @@ public interface ItemDao {
     int delete(ItemEntity item);
 
     @Query("SELECT * FROM items WHERE trip_id = :tripId ORDER BY id ASC")
-    List<ItemEntity> getItemsByTripId(int tripId);
+    List<ItemEntity> getItemsByTripId(long tripId);
 }
